Apply PR #14953 patch
