+++ config/environments/patches/database_configuration.rb
@@ -3,33 +3,8 @@
 # password fields as strings without ERB.
 
 require 'rails/application/configuration'
+require 'extensions/miq_db_config'
+
 Rails::Application::Configuration.module_eval do
-  prepend Module.new {
-    def database_configuration
-      path = paths["config/database"].existent.first
-      yaml = Pathname.new(path) if path
-
-      if yaml && yaml.exist?
-        require "yaml"
-        require "erb"
-
-        data = yaml.read
-        data = ERB.new(data).result if !Rails.env.production? || ENV['ERB_IN_CONFIG']
-
-        begin
-          data = YAML.load(data) || {}
-        rescue Psych::SyntaxError => e
-          raise "YAML syntax error occurred while parsing #{paths["config/database"].first}. " \
-                "Please note that YAML must be consistently indented using spaces. Tabs are not allowed. " \
-                "Error: #{e.message}"
-        rescue => e
-          raise e, "Cannot load `Rails.application.database_configuration`:\n#{e.message}", e.backtrace
-        end
-
-        Vmdb::Settings.decrypt_passwords!(data)
-      else
-        super
-      end
-    end
-  }
+  prepend MiqDbConfig
 end
